Add Go operator (#7)
This is the 1st version of the operator written in Go. Using these operator, users can interact with the cluster. The operator consists of CustomResource, represented by CustomResourceDefinition and controller loop, also called reconciliation loop. Coordinators and data instances are created as StatefulSets. Each coordinator and data instance have two services attached to it, ClusterIP and NodePort. ClusterIP is used for internal communication between instances and NodePort for external access to the Bolt server. Controller tests are deleted since they weren't used. Setup job runs as Deployment. License and organization name should be user-provided in the form of Kubernetes Secret.